# addrman: Improve performance of Good