@jonatack 2021-09-14
Nice code reduction.  Did you see a difference in `./src/bench/bench_bitcoin -filter=AddrManGood`?