@jonatack 2021-09-14
Concept ACK. Benches are showing a huge speedup for me locally.

before
```
$ ./src/bench/bench_bitcoin -filter=AddrManGood

|               ns/op |                op/s |    err% |     total | benchmark
|--------------------:|--------------------:|--------:|----------:|:----------
|      801,906,990.00 |                1.25 |    5.7% |      4.26 | AddrManGood
|      814,159,543.00 |                1.23 |    6.4% |      4.03 | AddrManGood
|      775,065,136.00 |                1.29 |    3.5% |      3.82 | AddrManGood
|      795,855,140.00 |                1.26 |    6.8% |      3.98 | AddrManGood
|      763,767,446.00 |                1.31 |    1.7% |      3.93 | AddrManGood
|      778,016,971.00 |                1.29 |    2.9% |      4.03 | AddrManGood
|      839,720,704.00 |                1.19 |    8.6% |      4.22 | AddrManGood
|      777,639,920.00 |                1.29 |    1.8% |      3.92 | AddrManGood
|      788,152,208.00 |                1.27 |    3.8% |      3.89 | AddrManGood
|      792,925,660.00 |                1.26 |    3.2% |      3.99 | AddrManGood
```

after
```
|               ns/op |                op/s |    err% |     total | benchmark
|--------------------:|--------------------:|--------:|----------:|:----------
|        5,182,474.00 |              192.96 |   13.8% |      0.03 | AddrManGood
|        5,525,509.00 |              180.98 |   14.6% |      0.03 | AddrManGood
|        7,215,235.00 |              138.60 |    3.7% |      0.04 | AddrManGood
|        5,219,675.00 |              191.58 |    7.5% |      0.03 | AddrManGood
|        4,163,823.00 |              240.16 |    2.3% |      0.02 | AddrManGood
|        6,694,095.00 |              149.39 |   23.0% |      0.05 | AddrManGood
|        5,966,011.00 |              167.62 |   11.0% |      0.03 | AddrManGood
|        4,861,261.00 |              205.71 |    3.6% |      0.03 | AddrManGood
|        5,029,188.00 |              198.84 |   16.4% |      0.03 | AddrManGood
|        5,440,878.00 |              183.79 |   13.1% |      0.03 | AddrManGood
```