@mzumsande 2021-09-14
Thanks! I didn't know about this benchmark, will try to confirm locally and add the result to OP!