@naumenkogs 2021-09-15
Concept ACK.
Code also looks good, but I'm gonna ACK once you resolve all the pending suggestions from other reviewers. I agree with those suggestions.