@mzumsande 2021-09-15
Thanks for the reviews - I addressed all comments.
I added a commit to allow for a lower number of sources in the fuzz test as suggested by @MarcoFalke .

I also ran the benchmark and got similar results to @jonatack:

`./src/bench/bench_bitcoin -filter=AddrManGood`
before:
```
|               ns/op |                op/s |    err% |     total | benchmark
|--------------------:|--------------------:|--------:|----------:|:----------
|      840,411,590.00 |                1.19 |    3.3% |      4.23 | `AddrManGood`
|      871,339,127.00 |                1.15 |    0.1% |      4.35 | `AddrManGood`
|      880,275,738.00 |                1.14 |    0.8% |      4.45 | `AddrManGood`
|      874,042,056.00 |                1.14 |    1.7% |      4.34 | `AddrManGood`
|      877,817,450.00 |                1.14 |    2.1% |      4.34 | `AddrManGood`
```
after
```
|               ns/op |                op/s |    err% |     total | benchmark
|--------------------:|--------------------:|--------:|----------:|:----------
|        4,889,438.00 |              204.52 |    0.7% |      0.02 | `AddrManGood`
|        4,831,691.00 |              206.97 |    1.5% |      0.02 | `AddrManGood`
|        4,847,879.00 |              206.28 |    1.0% |      0.02 | `AddrManGood`
|        4,745,268.00 |              210.74 |    0.8% |      0.02 | `AddrManGood`
|        4,932,979.00 |              202.72 |    1.3% |      0.02 | `AddrManGood`
```