@jnewbery 2021-09-16
ACK 57ce20307e

Using the public interface to fuzz addrman is a much better approach. As well as being fewer lines of code, the fuzz test is more correct now. The previous code reached into the internals and made changes to private data which invalidated addrman's invariants. For example, entries were placed into the tried table without setting `nLastSuccess`, which would cause `check()` to fail:

https://github.com/bitcoin/bitcoin/blob/87780dfee8c82ea09b84e74d331db2ad6f9e75ab/src/addrman.cpp#L774-L775

I think the next step is to move `RandAddr()` and `Fill()` out of `CAddrManDeterministic` so that they don't have access to `CAddrMan`'s private data and methods.