@laanwj 2021-09-20
This is fantastic, faster *and* less code.
Code review ACK 57ce20307e604530f78ef4f0f8d9fb94f80ca81b